requirements.txt
tensorflow>=2.5.0
pillow>=8.3.0
matplotlib>=3.4.2
numpy>=1.19.5
utils>=1.0.1
scikit-learn>=0.24.2
tensorboard>=2.5.0
opencv-python>=4.5.2.54